v1.2.2 Interim Update
This fixed a number of bugs, adds a number of new features including color-coded notation for lightning bolts and missed notes, and patches for the Z update.  

Some preliminary support for spin & spiral is included in this version and, but scores are not yet supported. It is not yet complete.